Switches to router example and drops tinymem dependency
This switches to a more sophisticated, but still simple router example.
This showcases how you can access multiple parts of the HTTP stream and
also serve locally.

This also drops the tinymem dependency because it is only used for a
single function and also adds alloc exports unneeded for this ABI.

Signed-off-by: Adrian Cole <adrian@tetrate.io>