Makes log levels begin with zero
While not specified, component-model aka wit generates enums that begin
with the ordinal zero and there's no way to change that. This defers to
zero indexing for log levels, which is less nice than -1, but better
for the ecosystem.

The alternative would be to use documentation to declare all the log
levels because we don't have flexibility in the "spec" (component-model)
or the generator for it (wit-bindgen). I think this is worse overall vs
someone accidentally choosing debug because it was number zero.

Signed-off-by: Adrian Cole <adrian@tetrate.io>